refactor: move resource parser contracts to accelerator package
