+++ internal/accelerator/plugin/amd_gpu.go
@@ -12,7 +12,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
-	resourceview "github.com/neutree-ai/neutree/internal/resource"
+	"github.com/neutree-ai/neutree/internal/accelerator/resourceparser"
 	"github.com/neutree-ai/neutree/pkg/command"
 	"github.com/neutree-ai/neutree/pkg/command_runner"
 )
@@ -169,7 +169,7 @@ func (p *AMDGPUAcceleratorPlugin) GetResourceConverter() ResourceConverter {
 	return NewAMDGPUConverter()
 }
 
-func (p *AMDGPUAcceleratorPlugin) GetResourceParser() resourceview.ResourceParser {
+func (p *AMDGPUAcceleratorPlugin) GetResourceParser() resourceparser.ResourceParser {
 	return &AMDGPUResourceParser{}
 }
