+++ internal/accelerator/plugin/gpu.go
@@ -12,7 +12,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
-	resourceview "github.com/neutree-ai/neutree/internal/resource"
+	"github.com/neutree-ai/neutree/internal/accelerator/resourceparser"
 	"github.com/neutree-ai/neutree/pkg/command"
 	"github.com/neutree-ai/neutree/pkg/command_runner"
 )
@@ -167,6 +167,6 @@ func (p *GPUAcceleratorPlugin) GetResourceConverter() ResourceConverter {
 	return NewGPUConverter()
 }
 
-func (p *GPUAcceleratorPlugin) GetResourceParser() resourceview.ResourceParser {
+func (p *GPUAcceleratorPlugin) GetResourceParser() resourceparser.ResourceParser {
 	return &GPUResourceParser{}
 }
