+++ internal/resource/resource_client.go
@@ -14,6 +14,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
+	"github.com/neutree-ai/neutree/internal/accelerator/resourceparser"
 	"github.com/neutree-ai/neutree/internal/ray/dashboard"
 )
 
@@ -22,66 +23,14 @@ type ResourceClient interface {
 	ListEndpointInstances(ctx context.Context, opts ListEndpointInstancesOptions) ([]EndpointInstanceResource, error)
 }
 
-// ResourceParser handles the standard resource semantics for an accelerator.
-// Virtualized Kubernetes resources can extend this interface with the optional
-// KubernetesVirtualization* parser interfaces below.
-type ResourceParser interface {
-	ParseFromRay(resource map[string]float64) (*v1.ResourceInfo, error)
-	ParseFromKubernetes(resource map[corev1.ResourceName]k8sresource.Quantity, labels map[string]string) (*v1.ResourceInfo, error)
-}
-
-type KubernetesNodeResourceContext struct {
-	NodeName             string
-	AllocatableResources map[corev1.ResourceName]k8sresource.Quantity
-	AvailableResources   map[corev1.ResourceName]k8sresource.Quantity
-	Labels               map[string]string
-	Annotations          map[string]string
-	Pods                 []KubernetesPodResourceContext
-}
-
-type KubernetesPodResourceContext struct {
-	Namespace   string
-	Name        string
-	UID         string
-	NodeName    string
-	Labels      map[string]string
-	Requests    map[corev1.ResourceName]k8sresource.Quantity
-	Limits      map[corev1.ResourceName]k8sresource.Quantity
-	Annotations map[string]string
-}
-
-type KubernetesEndpointNodeResourceContext struct {
-	Name        string
-	Labels      map[string]string
-	Annotations map[string]string
-}
-
-type KubernetesEndpointResourceContext struct {
-	EndpointName string
-	Namespace    string
-	Pods         []KubernetesPodResourceContext
-	Nodes        map[string]KubernetesEndpointNodeResourceContext
-}
-
-type KubernetesResourceParseResult struct {
-	Allocatable         *v1.ResourceInfo
-	Available           *v1.ResourceInfo
-	Devices             []*v1.DeviceResource
-	AcceleratorMetadata map[v1.AcceleratorType]*v1.AcceleratorMetadata
-}
-
-// KubernetesVirtualizationResourceParser parses a node only when its labels,
-// annotations, and Pod allocations match the parser's virtualization backend.
-// The matched flag deliberately separates "not my node" from an empty result.
-type KubernetesVirtualizationResourceParser interface {
-	ParseKubernetesVirtualizationNode(input KubernetesNodeResourceContext) (*KubernetesResourceParseResult, bool, error)
-}
-
-// KubernetesVirtualizationEndpointResourceParser converts backend-specific Pod
-// allocation annotations into Neutree Endpoint replica resource semantics.
-type KubernetesVirtualizationEndpointResourceParser interface {
-	ParseKubernetesVirtualizationEndpoint(input KubernetesEndpointResourceContext) ([]EndpointInstanceResource, bool, error)
-}
+type ResourceParser = resourceparser.ResourceParser
+type KubernetesNodeResourceContext = resourceparser.KubernetesNodeResourceContext
+type KubernetesPodResourceContext = resourceparser.KubernetesPodResourceContext
+type KubernetesEndpointNodeResourceContext = resourceparser.KubernetesEndpointNodeResourceContext
+type KubernetesEndpointResourceContext = resourceparser.KubernetesEndpointResourceContext
+type KubernetesResourceParseResult = resourceparser.KubernetesResourceParseResult
+type KubernetesVirtualizationResourceParser = resourceparser.KubernetesVirtualizationResourceParser
+type KubernetesVirtualizationEndpointResourceParser = resourceparser.KubernetesVirtualizationEndpointResourceParser
 
 const BytesPerGiB = 1024 * 1024 * 1024
 
@@ -102,12 +51,7 @@ type ResourceNode struct {
 	AcceleratorMetadata map[v1.AcceleratorType]*v1.AcceleratorMetadata
 }
 
-type EndpointInstanceResource struct {
-	InstanceID string
-	ReplicaID  string
-	NodeID     string
-	Devices    []v1.DeviceAllocation
-}
+type EndpointInstanceResource = resourceparser.EndpointInstanceResource
 
 type K8sResourceClient struct {
 	client  client.Client
